supervisor/controller/requirements.txt
paho-mqtt
python-dotenv
influxdb-client
python-dotenv